+++ packages/coast_audio/lib/src/node/abstract/single_in_out_node.dart
@@ -1,0 +1,15 @@
+import 'package:coast_audio/coast_audio.dart';
+
+mixin SingleInNodeMixin on AudioNode {
+  @override
+  List<AudioInputBus> get inputs => [inputBus];
+
+  AudioInputBus get inputBus;
+}
+
+mixin SingleOutNodeMixin on AudioNode {
+  @override
+  List<AudioOutputBus> get outputs => [outputBus];
+
+  AudioOutputBus get outputBus;
+}