+++ packages/coast_audio/lib/src/node/bus/audio_read_result.dart
@@ -1,0 +1,8 @@
+class AudioReadResult {
+  const AudioReadResult({
+    required this.frameCount,
+    required this.isEnd,
+  });
+  final int frameCount;
+  final bool isEnd;
+}