+++ packages/coast_audio/lib/src/ring_buffer/frame_ring_buffer.dart
@@ -1,61 +1,56 @@
 import 'package:coast_audio/coast_audio.dart';
 
 class FrameRingBuffer extends SyncDisposable {
+  static final _finalizer = Finalizer<SyncDisposable>((d) => d.dispose());
+
   FrameRingBuffer({
-    required int frames,
-    required AudioFormat format,
+    required this.capacity,
+    required this.format,
     Memory? memory,
-  }) : _buffer = AllocatedAudioFrames(
-          length: frames,
-          format: format,
-          fillZero: false,
+  }) : _ringBuffer = RingBuffer(
+          capacity: format.bytesPerFrame * capacity,
           memory: memory,
         ) {
-    _ringBuffer = RingBuffer(
-      capacity: _buffer.sizeInBytes,
-      pBuffer: _rawBuffer.pBuffer.cast(),
-    );
+    _finalizer.attach(this, _ringBuffer, detach: this);
   }
 
-  final AllocatedAudioFrames _buffer;
-  late final AudioBuffer _rawBuffer = _buffer.lock();
   late final RingBuffer _ringBuffer;
 
-  bool _isDisposed = false;
+  final AudioFormat format;
+
+  final int capacity;
+
+  int get length => _ringBuffer.length ~/ format.bytesPerFrame;
 
   @override
-  bool get isDisposed => _isDisposed;
-
-  AudioFormat get format => _buffer.format;
-
-  int get capacity => _buffer.sizeInFrames;
-
-  int get length => _ringBuffer.length ~/ _buffer.format.bytesPerFrame;
+  bool get isDisposed => _ringBuffer.isDisposed;
 
   int write(AudioBuffer buffer) {
-    return _ringBuffer.write(buffer.pBuffer.cast(), 0, buffer.sizeInBytes) ~/ buffer.format.bytesPerFrame;
+    throwIfNotAvailable();
+    assert(buffer.format.isSameFormat(format));
+    return _ringBuffer.write(buffer.pBuffer.cast(), 0, buffer.sizeInBytes) ~/ format.bytesPerFrame;
   }
 
   int read(AudioBuffer buffer, {bool advance = true}) {
-    return _ringBuffer.read(buffer.pBuffer.cast(), 0, buffer.sizeInBytes, advance: advance) ~/ buffer.format.bytesPerFrame;
+    throwIfNotAvailable();
+    assert(buffer.format.isSameFormat(format));
+    return _ringBuffer.read(buffer.pBuffer.cast(), 0, buffer.sizeInBytes, advance: advance) ~/ format.bytesPerFrame;
   }
 
   int copyTo(FrameRingBuffer buffer, {required bool advance}) {
-    return _ringBuffer.copyTo(buffer._ringBuffer, advance: advance);
+    throwIfNotAvailable();
+    assert(buffer.format.isSameFormat(format));
+    return _ringBuffer.copyTo(buffer._ringBuffer, advance: advance) ~/ format.bytesPerFrame;
   }
 
   void clear() {
+    throwIfNotAvailable();
     _ringBuffer.clear();
   }
 
   @override
   void dispose() {
-    if (_isDisposed) {
-      return;
-    }
-    _isDisposed = true;
-    _buffer
-      ..unlock()
-      ..dispose();
+    _ringBuffer.dispose();
+    _finalizer.detach(this);
   }
 }